requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
spacy
scikit-learn
python-dotenv
requests
groq